dev - reformulação do código, simplicando a execução do arquivo principal.
